Re-organize mcc src directory
