+++ gs/frontend/mcc/src/common/Logs.tsx
