+++ gs/frontend/mcc/src/telemetry/TelemetryData.tsx
